pyproject.toml
[build-system]
requires = ["poetry-core==1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "GAIA"
version = "0.1.0"
description = "Exoplanet classification system with an explainable ANN model"
authors = ["Krzysztof Gonet <36476874+Savaed@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
aiohttp = "^3.8.5"
astropy = "^5.3.3"
duckdb = "^0.9.0"
hydra-core = "^1.3.2"
loguru = "^0.7.2"
numpy = "^1.26.0"
pandas = "^2.1.1"
pydantic = "^2.4.2"
rich = "^13.6.0"
scipy = "^1.11.3"
google-cloud-storage = "^2.11.0"
google-cloud-dataproc = "^5.6.0"
google-cloud-service-usage = "^1.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.ml.dependencies]
tensorflow = "^2.14.0"

[tool.poetry.group.spark.dependencies]
pyspark = "3.4"

[tool.poetry.group.ui.dependencies]
dash = "^2.13.0"
fakeredis = "^2.19.0"
plotly = "^5.17.0"

[tool.poetry.group.dev.dependencies]
covdefaults = "^2.3.0"
coverage = "^7.3.1"
hypothesis = "^6.87.1"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"

[tool.pytest.ini_options]
pythonpath = "."

[tool.coverage.run]
plugins = ["covdefaults"]
omit = [
  "**/conftest.py",
  "**/progress.py",
  "**/log.py",
  "**/model_selection.py",
]

[tool.coverage.report]
exclude_lines = ["def __repr__", "def __str__"]

[tool.ruff]
select = ["E", "F", "N", "I", "UP"]
# Exception name should be named with an Error suffix
ignore = ["N818"]
line-length = 100
target-version = "py311"

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.mypy]
mypy_path = "."
explicit_package_bases = true

# The following config is equivalent to --strict (as of mypy 1.0)
warn_unused_configs = true
disallow_any_generics = true
# disallow_subclassing_any = true
# disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
# disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
# no_implicit_reexport = true
strict_equality = true
extra_checks = true

# TODO: Remove after providing type hints in tests
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.poetry.scripts]
download-data = "gaia.scripts.download_data:main_wrapper"
preprocess-data = "gaia.scripts.preprocess_data:main_wrapper"
run-dashboard = "gaia.scripts.run_dashboard:main"
create-features = "gaia.scripts.submit_spark_create_features_job.main_wrapper"
